+++ tests/intergration/test_convex_api.py
@@ -5,7 +5,10 @@
 """
 import pytest
 import secrets
-from eth_utils import remove_0x_prefix
+from eth_utils import (
+    remove_0x_prefix,
+    add_0x_prefix
+)
 
 from convex_api.account import Account
 from convex_api.convex_api import ConvexAPI
@@ -15,6 +18,7 @@ from convex_api.exceptions import (
 )
 
 TEST_FUNDING_AMOUNT = 8888888
+
 
 def test_convex_api_language_setup(convex_url):
     convex = ConvexAPI(convex_url)
@@ -69,7 +73,6 @@ def test_convex_api_send_basic_scrypt(convex_url, test_account):
     assert 'value' in result
     assert(result['value'] == [2, 3, 4, 5, 6])
 
-
 def test_convex_api_get_balance_no_funds(convex_url):
     convex = ConvexAPI(convex_url)
     account = Account.create_new()
@@ -113,7 +116,7 @@ def test_convex_api_call(convex_url):
     request_amount = convex.request_funds(amount, account)
     result = convex.send(deploy_storage, account)
     assert(result['value'])
-    contract_address = result['value']
+    contract_address = add_0x_prefix(result['value'])
     test_number = secrets.randbelow(1000)
     call_set_result = convex.send(f'(call storage-example(set {test_number}))', account)
     assert(call_set_result['value'] == test_number)
@@ -142,7 +145,6 @@ def test_convex_api_call(convex_url):
     #address = convex.get_address(contract_address_api, account)
     assert(address == contract_address )
 
-
 def test_convex_api_transfer(convex_url):
     convex = ConvexAPI(convex_url)
     account_from = Account.create_new()
@@ -156,18 +158,16 @@ def test_convex_api_transfer(convex_url):
     balance_to = convex.get_balance(account_to)
     assert(balance_to == amount / 2)
 
-
-
 def test_convex_api_query_lisp(convex_url, test_account):
     convex = ConvexAPI(convex_url)
     result = convex.query(f'(address "{test_account.address_api}")', test_account)
     assert(result)
     # return value is the address as a checksum
-    assert(result['value'] == test_account.address_checksum)
+    assert(add_0x_prefix(result['value']) == test_account.address_checksum)
 
 def test_convex_api_query_scrypt(convex_url, test_account):
     convex = ConvexAPI(convex_url, ConvexAPI.LANGUAGE_SCRYPT)
     result = convex.query(f'address("{test_account.address_api}")', test_account)
     assert(result)
     # return value is the address as a checksum
-    assert(result['value'] == test_account.address_checksum)
+    assert(add_0x_prefix(result['value']) == test_account.address_checksum)